+++ tests/pytest_tests/system_tests/test_artifacts/test_wandb_artifacts_full.py
@@ -204,7 +204,7 @@ def test_edit_after_add(wandb_init):
 def test_uploaded_artifacts_are_unstaged(wandb_init, tmp_path, monkeypatch):
     # Use a separate staging directory for the duration of this test.
     monkeypatch.setenv("WANDB_DATA_DIR", str(tmp_path))
-    staging_dir = Path(wandb_sdk.internal.artifacts.get_staging_dir())
+    staging_dir = Path(wandb_sdk.internal.artifact_saver.get_staging_dir())
 
     def dir_size():
         return sum(f.stat().st_size for f in staging_dir.rglob("*") if f.is_file())
